Clicking delete button shows confirm prompt before deleting

+++ public/script.js
@@ -1,10 +1,13 @@
 const deleteReview = (deleteButton) => {
-  const reviewId = deleteButton.target.getAttribute('data-review-id')
-  fetch(`/deletereview/${reviewId}`, {method: 'delete', credentials: 'include'})
-    .then(location.reload())
-    .catch((error) => {
-      console.error(error)
-    })
+  const confirmDelete = confirm('Are you sure you want to delete this review?')
+  if (confirmDelete) {
+    const reviewId = deleteButton.target.getAttribute('data-review-id')
+    fetch(`/deletereview/${reviewId}`, {method: 'delete', credentials: 'include'})
+      .then(location.reload())
+      .catch((error) => {
+        console.error(error)
+      })
+  }
 }
 
 document.querySelectorAll('.delete').forEach((deleteButton) => {
